Re-factored the `codebase`
