--- app/src/main/java/com/aritra/notify/components/actions/NoList.kt
@@ -1,49 +1,0 @@
-package com.aritra.notify.components.actions
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.aritra.notify.R
-
-@Composable
-fun NoList(contentDescription: String, message: String) {
-    Column(
-        Modifier.fillMaxSize(), verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Image(
-            modifier = Modifier.fillMaxWidth(),
-            painter = painterResource(id = R.drawable.no_list),
-            contentDescription = "null"
-        )
-        Spacer(modifier = Modifier.height(20.dp))
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = contentDescription,
-            textAlign = TextAlign.Center,
-            fontFamily = FontFamily(
-                Font(R.font.poppins_medium)
-            )
-        )
-        Spacer(modifier = Modifier.height(10.dp))
-        Text(
-            modifier = Modifier.fillMaxWidth(),
-            text = message,
-            textAlign = TextAlign.Center,
-            fontFamily = FontFamily(Font(R.font.poppins_light))
-        )
-    }
-}
