+++ app/src/main/java/com/aritra/notify/ui/screens/notes/homeScreen/NoteScreen.kt
@@ -2,10 +2,14 @@
 
 package com.aritra.notify.ui.screens.notes.homeScreen
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -33,13 +37,16 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.aritra.notify.R
 import com.aritra.notify.components.actions.BackPressHandler
 import com.aritra.notify.components.actions.LayoutToggleButton
-import com.aritra.notify.components.actions.NoList
 import com.aritra.notify.components.note.GridNoteCard
 import com.aritra.notify.components.note.NotesCard
 
@@ -146,5 +153,35 @@ fun NoteScreen(
                 }
             }
         }
+    }
+}
+
+@Composable
+fun NoList(contentDescription: String, message: String) {
+    Column(
+        Modifier.fillMaxSize(), verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Image(
+            modifier = Modifier.fillMaxWidth(),
+            painter = painterResource(id = R.drawable.no_list),
+            contentDescription = "null"
+        )
+        Spacer(modifier = Modifier.height(20.dp))
+        Text(
+            modifier = Modifier.fillMaxWidth(),
+            text = contentDescription,
+            textAlign = TextAlign.Center,
+            fontFamily = FontFamily(
+                Font(R.font.poppins_medium)
+            )
+        )
+        Spacer(modifier = Modifier.height(10.dp))
+        Text(
+            modifier = Modifier.fillMaxWidth(),
+            text = message,
+            textAlign = TextAlign.Center,
+            fontFamily = FontFamily(Font(R.font.poppins_light))
+        )
     }
 }
